Fix ensuring that the buffer is fontified
Use `font-lock-ensure'.  It was created for that purpose and exists
since Emacs 25.1.  Since we already require at least Emacs 26.1,
there is no need to support Emacs releases that lack that function.

This function checks whether it is appropriate to do font-locking
and if that is not so, it does noting.

Remove code from `htmlize-file', whose purpose it was to prevent only
parts of the buffer being fontified early on, with the expectations
that the complete buffer would be fontified later on.  Unfortunately
those fontification-prevention bindings also wrapped the code that
is responsible for the full fontification.

Start to always show "Fontifing BUFFER..." and "Htmlizing BUFFER..."
messages.  Previously only the former was shown and only if the buffer
was above a certain threshold.  The purpose of that was to ensure that
the user is aware that it is not htmlize's fault if the process takes
a while.  Now we show the progress message because that is the right
thing to do even if it doesn't take particularly long.